Move knowledge group into its own package
