+++ Packages/EurofurenceApplication/Sources/EurofurenceApplication/Content Router/Routes/Knowledge/KnowledgeGroupContentRoute.swift
@@ -1,4 +1,5 @@
 import ComponentBase
+import KnowledgeGroupComponent
 
 public struct KnowledgeGroupContentRoute {
