+++ Packages/EurofurenceApplication/Sources/EurofurenceApplication/Content Router/Routes/Knowledge/ShowKnowledgeContentFromGroupListing.swift
@@ -1,5 +1,6 @@
 import ComponentBase
 import EurofurenceModel
+import KnowledgeGroupComponent
 
 public struct ShowKnowledgeContentFromGroupListing: KnowledgeGroupEntriesComponentDelegate {
