+++ Packages/EurofurenceApplication/Tests/EurofurenceApplicationTests/Application/Component Test Doubles/StubKnowledgeGroupEntriesComponentFactory.swift
@@ -1,5 +1,6 @@
 import EurofurenceApplication
 import EurofurenceModel
+import KnowledgeGroupComponent
 import UIKit
 
 class StubKnowledgeGroupEntriesComponentFactory: KnowledgeGroupEntriesComponentFactory {
