--- Packages/EurofurenceApplication/Tests/EurofurenceApplicationTests/Application/Components/Knowledge Group Entries/Presenter Tests/Test Doubles/CapturingKnowledgeGroupEntriesComponentDelegate.swift
@@ -1,13 +1,0 @@
-import EurofurenceApplication
-import EurofurenceModel
-import Foundation
-import XCTEurofurenceModel
-
-class CapturingKnowledgeGroupEntriesComponentDelegate: KnowledgeGroupEntriesComponentDelegate {
-
-    private(set) var selectedKnowledgeEntryIdentifier: KnowledgeEntryIdentifier?
-    func knowledgeGroupEntriesComponentDidSelectKnowledgeEntry(identifier: KnowledgeEntryIdentifier) {
-        selectedKnowledgeEntryIdentifier = identifier
-    }
-
-}
