+++ Packages/EurofurenceApplication/Tests/EurofurenceApplicationTests/Application/Components/Knowledge Groups/View Model Tests/DefaultKnowledgeGroupsViewModelFactoryTests.swift
@@ -1,5 +1,6 @@
 import EurofurenceApplication
 import EurofurenceModel
+import KnowledgeGroupComponent
 import XCTest
 import XCTEurofurenceModel
