+++ Packages/EurofurenceApplication/Tests/EurofurenceApplicationTests/Content Router/Route Tests/Knowledge/KnowledgeGroupContentRouteTests.swift
@@ -3,6 +3,7 @@ import EurofurenceModel
 import XCTComponentBase
 import XCTest
 import XCTEurofurenceModel
+import XCTKnowledgeGroupComponent
 
 class KnowledgeGroupContentRouteTests: XCTestCase {
