+++ Packages/EurofurenceApplication/Tests/EurofurenceApplicationTests/Test Utilities/Random Value Generation/KnowledgeListGroupViewModel+RandomValueProviding.swift
@@ -1,8 +1,7 @@
-import Darwin
 import EurofurenceApplication
-import EurofurenceModel
+import KnowledgeGroupComponent
 import TestUtilities
-import UIKit.UIImage
+import XCTKnowledgeGroupComponent
 
 extension KnowledgeListGroupViewModel: RandomValueProviding {
 
@@ -11,14 +10,6 @@ extension KnowledgeListGroupViewModel: RandomValueProviding {
                                            fontAwesomeCharacter: .random,
                                            groupDescription: .random,
                                            knowledgeEntries: .random)
-    }
-
-}
-
-extension KnowledgeListEntryViewModel: RandomValueProviding {
-
-    public static var random: KnowledgeListEntryViewModel {
-        return KnowledgeListEntryViewModel(title: .random)
     }
 
 }
