+++ Packages/EurofurenceComponents/Sources/KnowledgeGroupComponent/Module/KnowledgeGroupEntriesComponentDelegate.swift
