+++ Packages/EurofurenceComponents/Sources/KnowledgeGroupComponent/Scene/UIKit/KnowledgeListEntryTableViewCell.swift
