+++ Packages/EurofurenceComponents/Sources/KnowledgeGroupComponent/Scene/UIKit/StoryboardKnowledgeGroupEntriesSceneFactory.swift
@@ -1,3 +1,4 @@
+import ComponentBase
 import UIKit
 
 struct StoryboardKnowledgeGroupEntriesSceneFactory: KnowledgeGroupEntriesSceneFactory {
