+++ Packages/EurofurenceComponents/Sources/KnowledgeGroupComponent/View Model/DefaultKnowledgeGroupViewModelFactory.swift
