+++ Packages/EurofurenceComponents/Sources/KnowledgeGroupComponent/View Model/KnowledgeGroupViewModelFactory.swift
