+++ Packages/EurofurenceComponents/Sources/KnowledgeGroupComponent/View Model/KnowledgeListEntryViewModel.swift
@@ -1,0 +1,9 @@
+public struct KnowledgeListEntryViewModel: Equatable {
+
+    public var title: String
+    
+    public init(title: String) {
+        self.title = title
+    }
+
+}
