+++ Packages/EurofurenceComponents/Sources/XCTKnowledgeGroupComponent/KnowledgeListViewModel+RandomValueProviding.swift
@@ -1,0 +1,10 @@
+import KnowledgeGroupComponent
+import TestUtilities
+
+extension KnowledgeListEntryViewModel: RandomValueProviding {
+
+    public static var random: KnowledgeListEntryViewModel {
+        return KnowledgeListEntryViewModel(title: .random)
+    }
+
+}
