+++ Packages/EurofurenceComponents/Tests/KnowledgeGroupComponentTests/Presenter Tests/Test Doubles/CapturingKnowledgeGroupEntryScene.swift
@@ -1,5 +1,4 @@
-import EurofurenceApplication
-import EurofurenceModel
+import KnowledgeGroupComponent
 
 class CapturingKnowledgeGroupEntryScene: KnowledgeGroupEntryScene {
