+++ Packages/EurofurenceComponents/Tests/KnowledgeGroupComponentTests/Presenter Tests/Test Doubles/StubKnowledgeGroupEntriesSceneFactory.swift
@@ -1,5 +1,5 @@
-import EurofurenceApplication
 import EurofurenceModel
+import KnowledgeGroupComponent
 import UIKit
 
 class StubKnowledgeGroupEntriesSceneFactory: KnowledgeGroupEntriesSceneFactory {
