+++ Packages/EurofurenceComponents/Tests/KnowledgeGroupComponentTests/Presenter Tests/WhenBindingKnowledgeEntry_KnowledgeGroupEntriesPresenterShould.swift
@@ -1,5 +1,5 @@
-import EurofurenceApplication
 import EurofurenceModel
+import KnowledgeGroupComponent
 import XCTest
 
 class WhenBindingKnowledgeEntry_KnowledgeGroupEntriesPresenterShould: XCTestCase {
