+++ Packages/EurofurenceComponents/Tests/KnowledgeGroupComponentTests/Presenter Tests/WhenBuildingKnowledgeGroupEntriesComponent.swift
@@ -1,5 +1,5 @@
-import EurofurenceApplication
 import EurofurenceModel
+import KnowledgeGroupComponent
 import XCTest
 
 class WhenBuildingKnowledgeGroupEntriesComponent: XCTestCase {
