+++ Packages/EurofurenceComponents/Tests/KnowledgeGroupComponentTests/Presenter Tests/WhenSceneSelectsKnowledgeEntry_KnowledgeGroupEntriesPresenterShould.swift
@@ -1,5 +1,5 @@
-import EurofurenceApplication
 import EurofurenceModel
+import KnowledgeGroupComponent
 import XCTest
 
 class WhenSceneSelectsKnowledgeEntry_KnowledgeGroupEntriesPresenterShould: XCTestCase {
