+++ Packages/EurofurenceComponents/Tests/KnowledgeGroupComponentTests/View Model Tests/DefaultKnowledgeGroupEntriesInteractorShould.swift
@@ -1,5 +1,5 @@
-import EurofurenceApplication
 import EurofurenceModel
+import KnowledgeGroupComponent
 import XCTest
 import XCTEurofurenceModel
